requirements.txt
pyqtgraph~=0.11.0
numpy~=1.21.5
pip~=21.2.4
toml~=0.10.2
zlib~=1.2.12
wheel~=0.37.1
brotli~=1.0.9
openssl~=1.1.1n
cryptography~=3.4.8
py~=1.11.0
lxml~=4.8.0
tornado~=6.1
typing_extensions~=4.1.1
keyring~=23.4.0
pkginfo~=1.8.2
setuptools~=61.2.0
cython~=0.29.28
future~=0.18.2
docutils~=0.17.1
werkzeug~=2.0.3
flask~=1.1.2
pyserial~=3.5
pyqt5~=5.15.7